-add handle2 for authorization middleware -change how common.Language works -add trusted proxies config field -bug fix

+++ internal/api/http/controller/v1/mangas/chapter.go
@@ -6,8 +6,10 @@ import (
 	"manga-explorer/internal/common/status"
 	"manga-explorer/internal/domain/mangas/dto"
 	"manga-explorer/internal/domain/mangas/service"
+	"manga-explorer/internal/util"
 	"manga-explorer/internal/util/httputil"
 	"manga-explorer/internal/util/httputil/resp"
+	"manga-explorer/internal/util/opt"
 )
 
 func NewChapterController(chapterService service.IChapter) ChapterController {
@@ -28,6 +30,7 @@ func (m ChapterController) InsertChapterPage(ctx *gin.Context) {
 		resp.ErrorDetailed(ctx, stat, fieldsErr)
 		return
 	}
+
 	stat = m.chapterService.InsertChapterPage(&input)
 	resp.Conditional(ctx, stat, nil, nil)
 }
@@ -37,6 +40,7 @@ func (m ChapterController) DeleteChapterPage(ctx *gin.Context) {
 	stat, fieldsErr := httputil.BindJson(ctx, &pageInput)
 	if stat.IsError() {
 		resp.ErrorDetailed(ctx, stat, fieldsErr)
+		return
 	}
 
 	stat = m.chapterService.DeleteChapterPages(&pageInput)
@@ -63,6 +67,13 @@ func (m ChapterController) CreateChapterComments(ctx *gin.Context) {
 		return
 	}
 
+	claims, stat := common.GetClaims(ctx)
+	if stat.IsError() {
+		resp.Error(ctx, stat)
+		return
+	}
+	input.UserId = claims.UserId
+
 	stat = m.chapterService.CreateChapterComment(&input)
 	resp.Conditional(ctx, stat, nil, nil)
 }
@@ -84,6 +95,14 @@ func (m ChapterController) FindChapterComments(ctx *gin.Context) {
 		return
 	}
 
+	if !util.IsUUID(chapterId) {
+		resp.ErrorDetailed(ctx, status.Error(status.BAD_PARAMETER_ERROR), common.ParameterError{
+			Param: "chapter_id",
+			Error: "chapter_id parameter should be on uuid type",
+		})
+		return
+	}
+
 	comments, stat := m.chapterService.FindChapterComments(chapterId)
 	resp.Conditional(ctx, stat, comments, nil)
 }
@@ -93,6 +112,15 @@ func (m ChapterController) FindPageComments(ctx *gin.Context) {
 		resp.ErrorDetailed(ctx, status.Error(status.BAD_PARAMETER_ERROR), common.NewNotPresentParameter("page_id"))
 		return
 	}
+
+	if !util.IsUUID(pageId) {
+		resp.ErrorDetailed(ctx, status.Error(status.BAD_PARAMETER_ERROR), common.ParameterError{
+			Param: "pageId",
+			Error: "page_id parameter should be on uuid type",
+		})
+		return
+	}
+
 	comments, stat := m.chapterService.FindPageComments(pageId)
 	resp.Conditional(ctx, stat, comments, nil)
 }
@@ -116,6 +144,7 @@ func (m ChapterController) CreateChapter(ctx *gin.Context) {
 	stat = m.chapterService.CreateChapter(&input)
 	resp.Conditional(ctx, stat, nil, nil)
 }
+
 func (m ChapterController) DeleteChapter(ctx *gin.Context) {
 	chapterId := ctx.Param("chapter_id")
 	if len(chapterId) == 0 {
@@ -123,28 +152,73 @@ func (m ChapterController) DeleteChapter(ctx *gin.Context) {
 		return
 	}
 
+	if !util.IsUUID(chapterId) {
+		resp.ErrorDetailed(ctx, status.Error(status.BAD_PARAMETER_ERROR), common.ParameterError{
+			Param: "chapter_id",
+			Error: "chapter_id parameter should be on uuid type",
+		})
+		return
+	}
+
 	stat := m.chapterService.DeleteChapter(chapterId)
 	resp.Conditional(ctx, stat, nil, nil)
 }
-func (m ChapterController) FindChapterPages(ctx *gin.Context) {
+
+func (m ChapterController) FindChapterDetails(ctx *gin.Context) {
 	chapterId := ctx.Param("chapter_id")
 	if len(chapterId) == 0 {
 		resp.ErrorDetailed(ctx, status.Error(status.BAD_PARAMETER_ERROR), common.NewNotPresentParameter("chapter_id"))
 		return
 	}
 
-	// TODO: Add feature to save the chapter as history
+	if !util.IsUUID(chapterId) {
+		resp.ErrorDetailed(ctx, status.Error(status.BAD_PARAMETER_ERROR), common.FieldError{
+			Field: "chapter_id",
+			Error: "chapter_id field should be on uuid type",
+		})
+		return
+	}
 
-	pages, stat := m.chapterService.FindChapterPages(chapterId)
+	var userId opt.Optional[string]
+	claims, stat := common.GetClaims(ctx)
+	if stat.IsError() {
+		userId = opt.NullStr
+	} else {
+		userId = opt.New(claims.UserId)
+	}
+
+	pages, stat := m.chapterService.FindChapterDetails(chapterId, userId)
 	resp.Conditional(ctx, stat, pages, nil)
 }
-func (m ChapterController) FindVolumeChapters(ctx *gin.Context) {
+
+func (m ChapterController) FindVolumeDetails(ctx *gin.Context) {
 	volumeId := ctx.Param("volume_id")
 	if len(volumeId) == 0 {
 		resp.ErrorDetailed(ctx, status.Error(status.BAD_PARAMETER_ERROR), common.NewNotPresentParameter("volume_id"))
 		return
 	}
 
-	chapters, stat := m.chapterService.FindVolumeChapters(volumeId)
+	chapters, stat := m.chapterService.FindVolumeDetails(volumeId)
 	resp.Conditional(ctx, stat, chapters, nil)
+}
+
+func (m ChapterController) GetMangaChapterHistories(ctx *gin.Context) {
+	input := dto.MangaChapterHistoriesFindInput{}
+	input.ConstructURI(ctx)
+
+	stat, fieldErrors := httputil.BindQuery(ctx, &input)
+	if stat.IsError() {
+		resp.ErrorDetailed(ctx, stat, fieldErrors)
+		return
+	}
+
+	claims, stat := common.GetClaims(ctx)
+	if stat.IsError() {
+		resp.Error(ctx, stat)
+		return
+	}
+	input.UserId = claims.UserId
+
+	histories, page, stat := m.chapterService.FindMangaChapterHistories(&input)
+	resp.Conditional(ctx, stat, histories, page)
 }
